Backport ATS 8.0.0 to stretch-wikimedia
Use Clang instead of GCC to build against stretch. Starting with version
8.0.0, ATS requires C++17, which no GCC version in Stretch supports.
C++17 is however supported by Clang 5 and 6, and they are available in
stretch-backports. There is unfortunately no recent version of libc++
available in stretch-backports, but we have uploaded it to
stretch-wikimedia/main.

Disable WCCP support as it makes the package FTBFS with the
stretch-backports Clang toolchain and we do not need it.

Stop shipping man pages. Building them requires Sphinx >= 1.5.1, not in
Stretch.

Bug: T204232
Change-Id: I0225fa8879458ba2f7e6021a584787e8f47b0eb1